Implemented the back-selection
So now this is (almost) fully functional, meaning you can either select from the panel or select from the viewport, and it (almost) works. There is one case, where if you select same object from the panel and then from the viewport, there is like one frame delay. I'll deal with this later.